[Refactor] Refactor CCE list nodes
